Create MediocrePE.iso
MediocrePE is a WindowsPE containing more than 100 different Native and 3rd-Party programs and utilities to assist in the task of collecting information, diagnosing, and troubleshooting/repairing a wide range of PC issues. I began this lil project to assist me in my day-to-day duties working as a Computer Technician.  So most of what you see is tailored to my daily needs at work, and stylized to my personal tastes. Not only do I make use of it on a daily basis, but all of my co-workers do as well. Now I would just like to put it out in the world in the hopes that it might find it's way into the hands of someone whom might also benefit from it. If you've managed to find yourself still reading up to this point, then please know that I very much appreciate your time and consideration!

Cheers
NicMediocre